게시글과 댓글 수정 삭제 권한 부여

+++ app/controllers/foods_controller.rb
@@ -1,5 +1,8 @@
 class FoodsController < ApplicationController
-  def posts
+  before_action :login_check
+	skip_before_action :login_check, :only => [:posts, :posts_category, :show]
+
+	def posts
   	@posts=Post.all
 	end
 
@@ -19,6 +22,7 @@ class FoodsController < ApplicationController
 
   def show
   	@post=Post.find(params[:id])
+		@comment_writer = User.where(id: session[:user_id])[0]
 	end
 
   def write
@@ -26,6 +30,7 @@ class FoodsController < ApplicationController
 
   def write_complete
 		post=Post.new
+		post.user_id = session[:user_id]
 		post.category=params[:post_category]
 		post.title=params[:post_title]
 		post.content=params[:post_content]
@@ -40,6 +45,10 @@ class FoodsController < ApplicationController
 
   def edit
   	@post=Post.find(params[:id])
+		if @post.user_id != session[:user_id]
+			flash[:alert] = "수정 권한이 없습니다."
+			redirect_to :back
+		end
 	end
 
   def edit_complete
@@ -52,15 +61,42 @@ class FoodsController < ApplicationController
       redirect_to "/foods/show/#{post.id}"
     else
       flash[:alert] = post.errors.values.flatten.join(' ')
-      redirect_to :back
-    end
-  end
+			redirect_to :back
+		end
+	end
 
   def delete_complete
   	post=Post.find(params[:id])
-		post.destroy
-		flash[:alert]="삭제되었습니다."
-		redirect_to "/"
+		if post.user_id == session[:user_id]
+			post.destroy
+			flash[:alert]="삭제되었습니다."
+			redirect_to "/"
+		else
+			flash[:alert]="삭제 권한이 없습니다."
+			redirect_to :back
+		end
+	end
+	
+	def write_comment_complete
+		comment = Comment.new
+		comment.user_id = session[:user_id]
+		comment.post_id = params[:post_id]
+		comment.content = params[:comment_content]
+		comment.save
+		
+		flash[:alert] = "새 댓글을 달았습니다."
+		redirect_to "/foods/show/#{comment.post.id}"
 	end
 
+	def delete_comment_complete
+		comment = Comment.find(params[:id])
+		if comment.user_id == session[:user_id]
+			comment.destroy
+			flash[:alert] = "댓글이 삭제되었습니다."
+			redirect_to "/foods/show/#{comment.post_id}"
+		else
+			flash[:alert] = "해당 댓글의 삭제 권한이 없습니다."
+			redirect_to :back
+		end
+	end
 end
